Moved code to components: added new journal folder w/JournalBlock
